Update to quire-11yt 1.0.0-rc.35

+++ _plugins/transforms/outputs/epub/index.js
@@ -1,11 +1,11 @@
-const fs = require('fs-extra')
-const manifestFactory = require('./manifest.js')
-const path = require('path')
-const sass = require('sass')
-const transform = require('./transform.js')
-const writer = require('./writer.js')
+import fs from 'fs-extra'
+import manifestFactory from './manifest.js'
+import path from 'node:path'
+import * as sass from 'sass'
+import transform from './transform.js'
+import writer from './writer.js'
 
-module.exports = (eleventyConfig, collections) => {
+export default (eleventyConfig, collections) => {
   const { outputDir } = eleventyConfig.globalData.config.epub
   const write = writer(outputDir)
 
@@ -34,8 +34,14 @@ module.exports = (eleventyConfig, collections) => {
     const assetDirsToCopy = ['fonts']
 
     assetDirsToCopy.forEach((name) => {
-      const source = path.join(eleventyConfig.dir.input, assetsDir, name)
+      const source = path.join(eleventyConfig.directoryAssignments.input, assetsDir, name)
       const dest = path.join(outputDir, assetsDir, name)
+
+      if (!fs.existsSync(source)) {
+        console.warn(`The asset directory ${source} does not exist.`)
+        return
+      }
+
       fs.copySync(source, dest)
     })
 
@@ -43,24 +49,68 @@ module.exports = (eleventyConfig, collections) => {
      * Copy styles
      */
     const sassOptions = {
-      loadPaths: [
-        path.resolve('node_modules')
+      api: 'modern-compiler',
+      loadPaths: [path.resolve('node_modules')],
+      silenceDeprecations: [
+        'color-functions',
+        'global-builtin',
+        'import',
+        'legacy-js-api',
+        'mixed-decls'
       ]
     }
-    const styles = sass.compile(path.resolve('content', assetsDir, 'styles', 'epub.scss'), sassOptions)
-    write(path.join(assetsDir, 'epub.css'), styles.css)
+
+    const stylesPath = path.resolve(eleventyConfig.directoryAssignments.input, assetsDir, 'styles', 'epub.scss')
+    if (fs.existsSync(stylesPath)) {
+      const styles = sass.compile(stylesPath, sassOptions)
+      write(path.join(assetsDir, 'epub.css'), styles.css)
+    }
 
     /**
      * Copy assets
      */
+
     const { assets } = eleventyConfig.globalData.epub
     const { url: coverUrl } = manifest.resources.find(({ rel }) => rel === 'cover-image')
     assets.push(coverUrl)
+
+    const isUrl = /https?:\/\//
+
+    // NB: `asset` contains POSIX-style paths or a URL
     for (const asset of assets) {
-      fs.copySync(
-        path.join(eleventyConfig.dir.output, asset),
-        path.join(outputDir, asset)
-      )
+      let assetDir
+      const destPath = path.posix.join(outputDir, asset)
+
+      // Fetch assets from content/_assets, otherwise use public or _site
+      switch (true) {
+        case isUrl.test(asset):
+          continue
+
+        case asset.split('/').at(0) === '_assets':
+          assetDir = eleventyConfig.directoryAssignments.input
+          break
+
+        case eleventyConfig.globalData.directoryConfig.publicDir !== false:
+          assetDir = eleventyConfig.globalData.directoryConfig.publicDir
+          break
+
+        default:
+          assetDir = eleventyConfig.directoryAssignments.output
+      }
+
+      const srcPath = path.posix.join(assetDir, asset)
+
+      if (!fs.existsSync(srcPath)) {
+        console.warn(`Asset ${srcPath} not present for copy, skipping...`)
+        continue
+      }
+
+      try {
+        fs.copySync(srcPath, destPath)
+      } catch (err) {
+        console.error(err)
+        process.exit(1)
+      }
     }
   })
 }
